pyproject.toml
[project]
name = "manitoba-grid-intelligence"
version = "0.1.0"
description = "Real-time power grid analytics platform for Manitoba"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Grid Intelligence Team"}
]
keywords = ["power-grid", "analytics", "data-engineering", "timescaledb", "kafka"]

dependencies = [
    "kafka-python>=2.0.2",
    "psycopg2-binary>=2.9.9",
    "pandas>=2.2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]
airflow = [
    "apache-airflow>=2.8.1",
    "apache-airflow-providers-postgres>=5.10.0",
    "apache-airflow-providers-apache-kafka>=1.3.0",
]
processing = [
    "faust-streaming>=0.10.0",
    "confluent-kafka>=2.3.0",
]

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
